Added rotation and options evaluation
Added a function to rotate the to-be-placed tile, both clockwise and counter-clockwise. This rotates both the image and its underlying data.
Added options evaluation. This function's use is twofold: first, it check whether the to-be-placed tile has any possible spots to be placed in, considering all rotations. If not, a new tile is selected. If so, these options are visualised, corresponding to the current rotation of the to-be-placed tile.