Tile placement reimplementation
Rewrote all functions to build the board grid structure (new rows and columns) and initialize the board.
Added functionality to make sure that, if a space for a new tile is available, it must comply with all its adjacent tiles, instead of only one.
Fixed bug where a placed tile would not be visible.
Added safety statement in rotation function.