Meeple placement visualisation
Meeples are positioned on the tile that they are placed on. A transparent overlay is generated to be an overlay of the actual tile such that, when multiple meeples are on the same tile, they can each be removed separately. This is a WIP and not applicable with only expansions 1, 2 and 3.
Fixed the Join_possession function, where, instead of adjacent material patches, the same material patch index would be selected and therefore incorrect possessions would be merged.