Colour indicator event
Added functionality to update the colour indicators of the player list.
Added a loop to add all current lobby connections to the initial lobby visualisation, instead of only the client player.
Added feed event when a new player joins the lobby, to add them to the player list. This does not work as intended yet.

Issues
New players who join the lobby are not correctly added to the player list.