Player list update fix
Fixed the event where a new player joins the lobby. This did not visualise correctly. It is now solved by having placeholders 'reserving' the space of potential new players, up to 6 players (the number of colours). Making the player list dynamically change size IS NOT WORKING (YET) and should be added later.