Added player left feed event
First work on implementing players leaving a lobby, and consequently cleaning up the data that they produced.
Also set up a feed event for when a new admin is assigned, for example when the current lobby admin leaves the lobby, or (later on, not implemented yet) selects another player to be admin of the lobby.